@@ -1,19 +1,23 @@
 import express from 'express';
 import { injectable, inject } from 'inversify';
-import { blackListTokensRepositoryDB } from '../repositories/black-list-tokens-repository-db';
+import { BlackListTokensRepositoryDB } from '../repositories/black-list-tokens-repository-db';
 import { UsersRepositoryDB } from '../repositories/users-repository-db';
-import { jwtPassService } from '../utils/jwt-pass-service';
+import { JwtPassService } from '../utils/jwt-pass-service';
 
 @injectable()
 export class CheckTokenService {
-  constructor(@inject(UsersRepositoryDB) protected usersRepositoryDB: UsersRepositoryDB) {}
+  constructor(
+    @inject(UsersRepositoryDB) protected usersRepositoryDB: UsersRepositoryDB,
+    @inject(BlackListTokensRepositoryDB) protected blackListTokensRepositoryDB: BlackListTokensRepositoryDB,
+    @inject(JwtPassService) protected jwtPassService: JwtPassService,
+  ) {}
 
   async noBlockToken(req: express.Request, res: express.Response, next: express.NextFunction) {
     const authHeader = req.headers['authorization'];
     if (authHeader) {
       const accessToken = authHeader!.split(' ')[1];
       if (accessToken) {
-        const verifyUser = jwtPassService.verifyJwt(accessToken);
+        const verifyUser = this.jwtPassService.verifyJwt(accessToken);
         if (verifyUser) {
           const user = await this.usersRepositoryDB.getUserById(verifyUser!.id!);
           if (user) {
@@ -38,7 +42,7 @@ export class CheckTokenService {
     if (authHeader) {
       const accessToken = authHeader!.split(' ')[1];
       if (accessToken) {
-        const verifyUser = jwtPassService.verifyJwt(accessToken);
+        const verifyUser = this.jwtPassService.verifyJwt(accessToken);
         if (verifyUser) {
           const user = await this.usersRepositoryDB.getUserById(verifyUser!.id!);
           if (user) {
@@ -60,10 +64,10 @@ export class CheckTokenService {
   async refreshToken(req: express.Request, res: express.Response, next: express.NextFunction) {
     const tokenRefresh = req.cookies.refreshToken;
     if (tokenRefresh) {
-      const verifyUser = jwtPassService.verifyJwt(tokenRefresh);
+      const verifyUser = this.jwtPassService.verifyJwt(tokenRefresh);
       const user = verifyUser && (await this.usersRepositoryDB.getUserById(verifyUser.id!));
       if (verifyUser && user) {
-        const isChecked = tokenRefresh && (await blackListTokensRepositoryDB.checkToken(tokenRefresh));
+        const isChecked = tokenRefresh && (await this.blackListTokensRepositoryDB.checkToken(tokenRefresh));
         if (!isChecked) {
           req.user = user;
           return next();
