@@ -1,15 +1,21 @@
+import 'reflect-metadata';
+import { injectable } from 'inversify';
 import { BlackList } from '../models/blackListModel';
 
-export const blackListTokensRepositoryDB = {
+@injectable()
+export class BlackListTokensRepositoryDB {
   async addToken(token: string) {
     try {
       return await BlackList.create({ tokenValue: token });
     } catch (err) {
       return `Db error: ${err}`;
     }
-  },
+  }
   async checkToken(token: string) {
-    const blackToken = await BlackList.findOne({ tokenValue: token }).exec();
-    return blackToken;
-  },
-};
+    try {
+      return await BlackList.findOne({ tokenValue: token }).exec();
+    } catch (err) {
+      return `Db error: ${err}`;
+    }
+  }
+}
